[feature/allow-for-empty-dashboard] Allow for no environment dashboard (#928)
* [feature/allow-for-empty-dashboard] Allow for no environment dashboard

In 1.5.9 or earlier you could have all (default) or some of the
environments on the dashboard. In 1.6.0 that changed to none (default) or
some of the environments. This change changes it back to 1.5.9 - (all)
default, some or now none is directly supported as a use case.

* missing api file

+++ backend/mr-db-sql/src/main/java/io/featurehub/db/services/UserStateSqlApi.java
@@ -65,7 +65,8 @@ public class UserStateSqlApi implements UserStateApi {
       return;
     }
 
-    if (environments == null || environments.getEnvironmentIds() == null || environments.getEnvironmentIds().isEmpty()) {
+    if (environments == null || environments.getEnvironmentIds() == null ||
+        (environments.getEnvironmentIds().isEmpty() && environments.getNoneSelected() != Boolean.TRUE)) {
       userStateFinder(currentPerson, application.getId(), UserState.HIDDEN_FEATURES).delete();
       return;
     }
